LL automation

+++ src/main/java/ti4/discord/interactions/buttons/handlers/unit/AssignHitsButtonHandlers.java
@@ -7,7 +7,7 @@ import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import org.apache.commons.lang3.StringUtils;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.Iron.IronLeadersHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.ashen.AshenUnitHandler;
-import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.Ponthous.PonthousUnitHandler;
+import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.Ponthous.*;
 import ti4.discord.interactions.routing.ButtonHandler;
 import ti4.game.Game;
 import ti4.game.Player;
@@ -61,6 +61,10 @@ class AssignHitsButtonHandlers {
                     if (remove) {
                         RemoveUnitService.removeUnit(event, tile, game, unit, state);
                     } else {
+                        if (PonthousAbilityHandler.offerLastStand(
+                                event, game, player, tile, holder, unit.unitKey(), state, assignHitsType)) {
+                            return;
+                        }
                         DestroyUnitService.destroyUnit(event, tile, game, unit, combat, state);
                         IronLeadersHandler.checkCommanderUnlockAfterCombat(game, tile, holder, assignHitsType);
                     }
